refactor(import-network-panel): move host find panel to a separate file
